spark-jobs/requirements.txt
psycopg2-binary==2.9.9  # Postgres 접속/업서트
